refactor(schematics): split setup schematic on multiple files
